+++ RealmSwift/Tests/RealmTests.swift
@@ -1903,3 +1903,120 @@ extension CancellationError: Equatable {
         true
     }
 }
+
+@available(macOS 12.0, watchOS 8.0, iOS 15.0, tvOS 15.0, macCatalyst 15.0, *)
+final class InMemoryLogger: Logger {
+    var logs: String = ""
+
+    init(level: LogLevel) {
+        super.init()
+        self.level = level
+    }
+
+    override func doLog(level: LogLevel, message: String) {
+        logs += "\(Date.now) \(level.logLevel) \(message)"
+    }
+
+    func clear() {
+        logs = ""
+    }
+}
+
+extension LogLevel {
+    var logLevel: String {
+        switch self {
+        case .off:
+            return "Off"
+        case .fatal:
+            return "Fatal"
+        case .error:
+            return "Error"
+        case .warn:
+            return "Warn"
+        case .info:
+            return "Info"
+        case .detail:
+            return "Details"
+        case .debug:
+            return "Debug"
+        case .trace:
+            return "Trace"
+        case .all:
+            return "All"
+        default:
+            return "unknown"
+        }
+    }
+}
+
+@available(macOS 12.0, watchOS 8.0, iOS 15.0, tvOS 15.0, macCatalyst 15.0, *)
+class LoggerTests: TestCase {
+    func testSetDefaultLogLevel() throws {
+        let inMemoryLogger = InMemoryLogger(level: .off)
+        Logger.setDefaultLogger(inMemoryLogger)
+
+        try autoreleasepool { _ = try Realm() }
+        XCTAssertTrue(inMemoryLogger.logs.isEmpty)
+
+        // Change default log threshold level, this should change the log level of the custom logger
+        Logger.logLevel = .all
+        try autoreleasepool { _ = try Realm() } // We should be getting logs after changing the log level
+        XCTAssertEqual(Logger.logLevel, .all)
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB:", options: .regularExpression))
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Trace DB:", options: .regularExpression))
+    }
+
+    func testCustomLogger() throws {
+        let inMemoryLogger = InMemoryLogger(level: .off)
+        Logger.setDefaultLogger(inMemoryLogger)
+
+        try autoreleasepool { _ = try Realm() }
+
+        XCTAssertTrue(inMemoryLogger.logs.isEmpty)
+
+        // Info
+        inMemoryLogger.level = .info
+        try autoreleasepool { _ = try Realm() }
+
+        XCTAssertTrue(!inMemoryLogger.logs.isEmpty)
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB:", options: .regularExpression))
+
+        // Debug
+        inMemoryLogger.clear()
+        inMemoryLogger.level = .trace
+        try autoreleasepool { _ = try Realm() }
+
+        XCTAssertTrue(!inMemoryLogger.logs.isEmpty)
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Trace DB:", options: .regularExpression))
+
+        // Info
+        inMemoryLogger.clear()
+        inMemoryLogger.level = .info
+        try autoreleasepool { _ = try Realm() }
+
+        XCTAssertTrue(!inMemoryLogger.logs.isEmpty)
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB:", options: .regularExpression))
+        XCTAssertNil(inMemoryLogger.logs.range(of: "Trace DB:", options: .regularExpression))
+    }
+
+    func testCustomLoggerLog() throws {
+        let inMemoryLogger = InMemoryLogger(level: .info)
+        Logger.setDefaultLogger(inMemoryLogger)
+        XCTAssertTrue(inMemoryLogger.logs.isEmpty)
+
+        try autoreleasepool { _ = try Realm() }
+
+        XCTAssertTrue(!inMemoryLogger.logs.isEmpty)
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB:", options: .regularExpression))
+
+        inMemoryLogger.log(level: .info, message: "Info DB: 'Database is broken'")
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB:", options: .regularExpression))
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB: 'Database is broken'", options: .regularExpression))
+
+        inMemoryLogger.log(level: .debug, message: "Debug DB: 'Database is good'")
+        // Should not be added to log, because log level is greater than the one set.
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB:", options: .regularExpression))
+        XCTAssertNotNil(inMemoryLogger.logs.range(of: "Info DB: 'Database is broken'", options: .regularExpression))
+        XCTAssertNil(inMemoryLogger.logs.range(of: "Debug DB: 'Database is good'", options: .regularExpression))
+    }
+}